apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
sqlalchemy==2.0.27
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
diffusers==0.26.3
torch==2.2.0
transformers==4.37.2
accelerate==0.27.2
google-genai
